+++ reflex_django_tests/test_reflex_mount_urls.py
@@ -2,4 +2,4 @@
 
 from reflex_django.urls import reflex_mount
 
-urlpatterns = [reflex_mount()]
+urlpatterns = list(reflex_mount())